Split reports/_sections.py into a sections/ package by domain
The section builders lived in a single ~3,750-line module. Move them into
gem.reports.sections, grouped by domain (match, economy, combat, vision),
with a small _shared.py for the three cross-group helpers (_RADIANT_COLORS,
_DIRE_COLORS, _ward_enemies_seen).

_sections.py is now a thin re-export shim so existing imports
(builder.py, tests) keep working unchanged. Each new module imports only
the names it uses.

Pure refactor: generated HTML is byte-identical (verified by SHA256 over a
full report from a 10-player synthetic match). ruff + mypy clean, 1534 fast
tests pass.
